+++ plugin-packs/postcss-preset-env/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changes to PostCSS Preset Env
 
-### Unreleased (patch)
+### Unreleased (minor)
 
+- Add support for array options, e.g. `{ 'nesting-rules': ['auto', { noIsPseudoSelector: false }] }`
 - Updated [`@csstools/postcss-gradients-interpolation-method`](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-gradients-interpolation-method) to [`4.0.11`](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-gradients-interpolation-method/CHANGELOG.md#4011) (patch)
 
 ### 9.4.0